+++ deployments/docker/images/inbox/Dockerfile
@@ -9,7 +9,6 @@ RUN ssh-keygen -t rsa     -N '' -f /etc/ssh/ssh_host_rsa_key && \
     ssh-keygen -t dsa     -N '' -f /etc/ssh/ssh_host_dsa_key && \
     ssh-keygen -t ecdsa   -N '' -f /etc/ssh/ssh_host_ecdsa_key && \
     ssh-keygen -t ed25519 -N '' -f /etc/ssh/ssh_host_ed25519_key && \
-    useradd ega && \
     mkdir -p /usr/local/lib/ega && \
     echo '/usr/local/lib/ega' > /etc/ld.so.conf.d/ega.conf && \
     echo 'Welcome to Local EGA' > /ega/banner && \
@@ -18,19 +17,7 @@ RUN ssh-keygen -t rsa     -N '' -f /etc/ssh/ssh_host_rsa_key && \
     git clone https://github.com/NBISweden/LocalEGA-auth /root/ega-auth && \
     cd /root/ega-auth/src && \
     make install clean && \
-    ldconfig -v && \
-    chown lega:lega /ega/inbox && \
-    chmod 750 /ega/inbox && \
-    chmod g+s /ega/inbox
-
-# S3 File System
-RUN pip3.6 install boto3 && \
-    yum -y install automake fuse fuse-devel gcc-c++ git libcurl-devel libxml2-devel make openssl-devel && \
-    git clone https://github.com/s3fs-fuse/s3fs-fuse.git && \
-    cd s3fs-fuse && \
-    ./autogen.sh && \
-    ./configure && \
-    make && make install
+    ldconfig -v
 
 COPY pam.ega /etc/pam.d/ega
 COPY sshd_config /etc/ega/sshd_config
@@ -38,4 +25,3 @@ RUN cp /usr/sbin/sshd /usr/sbin/ega
 COPY entrypoint.sh /usr/local/bin/entrypoint.sh
 RUN chmod 755 /usr/local/bin/entrypoint.sh
 ENTRYPOINT ["entrypoint.sh"]
-
