+++ deployments/docker/images/vault/entrypoint.sh
@@ -6,16 +6,17 @@ set -e
 [[ -z "$MQ_INSTANCE" ]] && echo 'Environment MQ_INSTANCE is empty' 1>&2 && exit 1
 
 # Changing permissions
-echo "Changing permissions for /ega/vault and /ega/staging"
-chown lega:lega /ega/vault /ega/staging
-chmod 750 /ega/vault /ega/staging
-chmod g+s /ega/vault /ega/staging # setgid bit
+echo "Changing permissions for /ega/vault"
+chown lega:lega /ega/vault
+chmod 750 /ega/vault 
+chmod g+s /ega/vault # setgid bit
 
 echo "Waiting for Local Message Broker"
 until nc -4 --send-only ${MQ_INSTANCE} 5672 </dev/null &>/dev/null; do sleep 1; done
 
 echo "Starting the verifier"
-ega-verify &
+#exec gosu lega ega-verify
+exec ega-verify
 
-echo "Starting the vault listener"
-exec gosu lega ega-vault
+# echo "Starting the vault listener"
+# exec gosu lega ega-vault
