+++ extras/db.sql
@@ -13,18 +13,12 @@ CREATE EXTENSION pgcrypto;
 CREATE TABLE files (
         id             SERIAL, PRIMARY KEY(id), UNIQUE (id),
 	elixir_id      TEXT NOT NULL,
-	filename       TEXT NOT NULL,
-	enc_checksum   TEXT,
-	enc_checksum_algo hash_algo,
-	org_checksum   TEXT,
-	org_checksum_algo hash_algo,
+	inbox_path     TEXT NOT NULL,
 	status         status,
-	staging_name   TEXT,
+	vault_path     TEXT,
+	vault_filesize Integer,
 	stable_id      TEXT,
-	filepath       TEXT,
-	reenc_info     TEXT,
-	reenc_size     INTEGER,
-	reenc_checksum TEXT, -- sha256
+	header         TEXT, -- crypt4gh
 	created_at     TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT clock_timestamp(),
 	last_modified  TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT clock_timestamp()
 );
@@ -38,7 +32,7 @@ CREATE FUNCTION insert_file(filename    files.filename%TYPE,
     DECLARE
         file_id files.id%TYPE;
     BEGIN
-	INSERT INTO files (filename,elixir_id,stable_id,status)
+	INSERT INTO files (inbox_path,elixir_id,stable_id,status)
 	VALUES(filename,eid,stable_id,status) RETURNING files.id
 	INTO file_id;
 	RETURN file_id;
