+++ lega/conf/defaults.ini
@@ -5,19 +5,30 @@ log = nope
 [ingestion]
 do_checksum = True
 
+[quality_control]
+keyserver_endpoint = https://ega_keys:9000/retrieve/%s/private
+
 [inbox]
 location = /ega/inbox/%s
 mode = 2750
-#driver = file, s3
-
-[s3]
-url = http://ega_s3:9000
-region = lega
-access_key = LEGA-ACCESS-KEY
-secret_key = LEGA-SECRET-KEY
 
 [vault]
+###########################
+# Back by POSIX fs
+###########################
 location = /ega/vault
+mode = 2750
+driver = lega.vault.FileStorage
+
+###########################
+# Backed by S3
+###########################
+# driver = lega.vault.S3Storage
+url = http://ega_vault_s3:9000
+region = lega
+bucket = lega
+access_key = LEGA-VAULT-ACCESS-KEY
+secret_key = LEGA-VAULT-SECRET-KEY
 
 ## Connecting to Local Broker
 [broker]
